+++ src/aws/s3_scaleway.tf
@@ -48,7 +48,7 @@ resource "scaleway_object_bucket_policy" "policy" {
           Principal = "*",
           Action = "s3:GetObject",
           Resource = [
-            "${scaleway_object_bucket.marsha_object_bucket.name}/scw/*",
+            "${scaleway_object_bucket.marsha_object_bucket.name}/vod/*",
             "${scaleway_object_bucket.marsha_object_bucket.name}/tmp/*"
           ]
         },