+++ src/backend/marsha/development/api.py
@@ -9,6 +9,7 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
 from marsha.core.models import Document, Video
+from marsha.core.storage.storage_class import video_storage
 from marsha.core.utils import time_utils
 
 
@@ -16,8 +17,8 @@ logger = logging.getLogger(__name__)
 
 
 @api_view(["POST"])
-def local_video_upload(request: HttpRequest, uuid=None):
-    """Endpoint to mock s3 video upload."""
+def dummy_video_upload(request: HttpRequest, uuid=None):
+    """Dummy endpoint to mock s3 video upload."""
     try:
         object_instance = Video.objects.get(id=uuid)
     except Video.DoesNotExist:
@@ -30,6 +31,18 @@ def local_video_upload(request: HttpRequest, uuid=None):
     )
 
     return Response({"success": True}, status=204)
+
+
+@api_view(["POST"])
+def local_video_upload(request: HttpRequest, uuid=None, stamp=None):
+    """Endpoint to mock s3 video upload in dev environment."""
+    uploaded_video_file = request.FILES["file"]
+
+    video_storage.save(
+        f"tmp/{uuid}/video/{stamp}",
+        uploaded_video_file,
+    )
+    return Response(status=204)
 
 
 @api_view(["POST"])