+++ src/frontend/apps/standalone_site/src/setupProxy.js
@@ -2,7 +2,7 @@ const { createProxyMiddleware } = require('http-proxy-middleware');
 
 module.exports = function (app) {
   app.use(
-    ['/api', '/account/api'],
+    ['/api', '/account/api', '/data/videos/'],
     createProxyMiddleware({
       target: process.env.REACT_APP_BACKEND_API_BASE_URL,
     }),